feat: add component discovery.relabel
this also refactors the existing relabeling components to reuse the logic for
rule blocks
