+++ src/components/ComponentEditor/components/DiscoveryRelabel.tsx
@@ -1,36 +1,31 @@
 import { FormAPI, InlineField, Input } from "@grafana/ui";
 import ReferenceMultiSelect from "../inputs/ReferenceMultiSelect";
 import { RelabelRules, transformRules } from "../common/RelabelRules";
+import ReferenceSelect from "../inputs/ReferenceSelect";
 
 const Component = ({ methods }: { methods: FormAPI<Record<string, any>> }) => {
   return (
     <>
       <InlineField
-        label="Forward to"
-        tooltip="Where to forward log entries after relabeling."
+        label="Targets"
+        tooltip="Targets to relabel"
         labelWidth={22}
         error="You must specify a list of destinations"
         invalid={!!methods.errors["targets"]}
       >
-        <ReferenceMultiSelect
-          name="forward_to"
-          exportName="LokiReceiver"
+        <ReferenceSelect
+          name="targets"
+          exportName="list(Target)"
           control={methods.control}
         />
       </InlineField>
-      <InlineField
-        label="Max cache size"
-        tooltip="The maximum number of elements to hold in the relabeling cache."
-        labelWidth={22}
-      >
-        <Input type="number" {...methods.register("max_cache_size")} />
-      </InlineField>
+
       <RelabelRules methods={methods} />
     </>
   );
 };
 
-const LokiRelabel = {
+const DiscoveryRelabel = {
   preTransform(data: Record<string, any>): Record<string, any> {
     return data;
   },
@@ -39,4 +34,4 @@ const LokiRelabel = {
   },
   Component,
 };
-export default LokiRelabel;
+export default DiscoveryRelabel;
