+++ src/components/ComponentEditor/components/LokiSourceFile.tsx
@@ -14,7 +14,7 @@ const Component = ({ methods }: { methods: FormAPI<Record<string, any>> }) => {
       >
         <ReferenceSelect
           name="targets"
-          exportName="list(FileTarget)"
+          exportName="list(Target)"
           control={methods.control}
         />
       </InlineField>
